backend/requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pdfplumber==0.11.4
groq
python-dotenv
playwright
beautifulsoup4
requests
